make common files symlinks + remove makefile
